add util functions to simapp, fix bank keeper test
