+++ simapp/utils.go
@@ -515,68 +515,3 @@ func GetSimulationLog(storeName string, sdr sdk.StoreDecoderRegistry, cdc *codec
 
 	return
 }
-
-// GenTx generates a signed mock transaction.
-func GenTx(msgs []sdk.Msg, accnums []uint64, seq []uint64, priv ...crypto.PrivKey) auth.StdTx {
-	// Make the transaction free
-	fee := auth.StdFee{
-		Amount: sdk.NewCoins(sdk.NewInt64Coin("foocoin", 0)),
-		Gas:    100000,
-	}
-
-	sigs := make([]auth.StdSignature, len(priv))
-	memo := "testmemotestmemo"
-
-	for i, p := range priv {
-		sig, err := p.Sign(auth.StdSignBytes(chainID, accnums[i], seq[i], fee, msgs, memo))
-		if err != nil {
-			panic(err)
-		}
-
-		sigs[i] = auth.StdSignature{
-			PubKey:    p.PubKey(),
-			Signature: sig,
-		}
-	}
-
-	return auth.NewStdTx(msgs, fee, sigs, memo)
-}
-
-// SignCheckDeliver checks a generated signed transaction and simulates a
-// block commitment with the given transaction. A test assertion is made using
-// the parameter 'expPass' against the result. A corresponding result is
-// returned.
-func SignCheckDeliver(
-	t *testing.T, cdc *codec.Codec, app *baseapp.BaseApp, header abci.Header, msgs []sdk.Msg,
-	accNums, seq []uint64, expSimPass, expPass bool, priv ...crypto.PrivKey,
-) sdk.Result {
-
-	tx := GenTx(msgs, accNums, seq, priv...)
-
-	txBytes, err := cdc.MarshalBinaryLengthPrefixed(tx)
-	require.Nil(t, err)
-
-	// Must simulate now as CheckTx doesn't run Msgs anymore
-	res := app.Simulate(txBytes, tx)
-
-	if expSimPass {
-		require.Equal(t, sdk.CodeOK, res.Code, res.Log)
-	} else {
-		require.NotEqual(t, sdk.CodeOK, res.Code, res.Log)
-	}
-
-	// Simulate a sending a transaction and committing a block
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-	res = app.Deliver(tx)
-
-	if expPass {
-		require.Equal(t, sdk.CodeOK, res.Code, res.Log)
-	} else {
-		require.NotEqual(t, sdk.CodeOK, res.Code, res.Log)
-	}
-
-	app.EndBlock(abci.RequestEndBlock{})
-	app.Commit()
-
-	return res
-}
